requirements.txt
streamlit
openai-whisper
sentence-transformers
chromadb
ffmpeg-python
torch